requirements.txt
django==5.0.6
groq==0.31.1
python-dotenv==1.0.1
pillow==10.4.0
pytesseract==0.3.10
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn==22.0.0
whitenoise==6.7.0
